UPDATED CODE
