+++ project/songFunctions.py
@@ -1,0 +1,1 @@
+from playsound import playsound #pip install playsound==1.2.2import waveimport numpy as npimport soundfile as sfimport ioimport structimport scipy.io as scipy#must take in numpy array#must take in int: sample ratedef waveSpeed(data, speed_amt):      #for use after the intial scipy.io.wavfile read()    print(f"number of channels = {data.shape[1]}")
